Tool policies (#263)
dbt Labs' policies dictate that we do not feed row-level data to LLMs in
our systems. There is no such restriction for typical users of dbt-mcp
because they have their own terms with LLM providers. However, when we
use dbt-mcp internally, we need to ensure that we don't send tool
results with row-level data to LLMs. This PR applies policies to each
tool to ensure we are always compliant.

I had to change the get_model_details tool to return raw SQL instead of
compiled SQL in order for it to be compliant with metadata only.