Oauth initial implementation (#311)
This PR is the initial implementation of oauth login for the MCP server
running locally. There are a few main components of this PR:

1. The initial oauth flow with dbt Platform login and redirect to
localhost.
2. Hosting of static HTML, CSS, and JS assets with a FastAPI server.
3. APIs to retrieve dbt Project information provided by that same
FastAPI server.
4. A frontend React UI to select and view dbt Projects.

These components work in concert. For example:

* When a user selects a project, that information is stored in a local
yaml file `mcp.json` to be automatically reused for the next time.
* When the browser window is closed, an API request is sent to the
FastAPI server to shut it down since it is no longer needed.
* After the login process is finished, the MCP server is configured
based on the user's selection and continues executing as it normally
would.

I tested this OAuth flow end-to-end, as well as tested the previous env
var-based configuration, to ensure I didn't break anything. This feature
is behind an undocumented env var while it is still being developed.