Handle Claude Desktop running multiple MCP server instances (#320)
Claude Desktop is starting multiple instances of the MCP server. I think
this behavior has always occurred, but now that we have OAuth, the issue
surfaces as poor UX when multiple OAuth flows are initiated at the same
time.

I reproduced this issue with a toy MCP server, so it has little to do
with our MCP implementation in this repo. I also asked a coworker to
reproduce this error, so it is not related to my specific installation
of Claude Desktop. I also experimented with a more native oauth
implementation with
[this](https://github.com/modelcontextprotocol/python-sdk?tab=readme-ov-file#authentication)
example, but that did not change Claude Desktop's behavior.

The fix in this PR creates a lock file so that when Claude Desktop
starts multiple dbt MCP instances, only the first completes the oauth
flow while latter instances are blocked until the oauth flow is
complete. When subsequent dbt MCP instances become unblocked from the
lock, they will read the dbt Platform Context from `mcp.yml` rather than
initializing another oauth flow. This effectively fixes the issue of
multiple OAuth flows being triggered. However, this strategy also has a
deeper reliance on `mcp.yml`. Once `mcp.yml` is set, the oauth flow will
not trigger, so in order to re-configure dbt MCP with oauth, this file
needs to be deleted out of band by the user.