Automatically refresh oauth token (#329)
This PR implements periodic oauth token refresh in the background. This
ensures that the user doesn't need to re-authenticate frequently.

Now, there are two places that require interacting with the dbt Platform
OAuth API: the FastAPI callback & the refresh token background process.
This meant a lot of the code had to be refactored so that it could be
shared between these two places as much as possible.

Furthermore, the configuration classes had to be changed to utilize
`HeadersProvider` rather than a static headers dictionary. This is
because the authorization headers are changed each time the oauth token
is refreshed. Using a service token or PAT is still supported through
the `StaticTokenProvider` class.

I tested this with `task client`. I ensured that the pre-existing
environment variable-based configuration continued to work as expected
for all dbt Platform toolsets. To test oauth token refresh, I used the
`MockRefreshStrategy` and set a low refresh interval (1 second). I then
watched the logs & `mcp.yml` update as expected with refreshed oauth
tokens.

Other notable changes include:
* Update to the app lifespan to initiate the refresh token background
process.
* Creation of `DbtPlatformContextManager` to manage the `mcp.yml` file.
* `RefreshStrategy` protocol and implementations to define when the
oauth token should be refreshed.
* `OAuthTokenProvider` where the token is refreshed in a background job.

Please let me know if there is any way I can make this PR easier to
review.