Address claude desktop re-triggering oauth on exit (#340)
Claude Desktop starts the MCP multiple times. Twice on start-up and once
again when you close the app. This was problematic because the oauth
flow would trigger eagerly when the dbt MCP server started. This PR
changes the behavior so that the oauth flow is triggered lazily, when
tool calls are made. This fixes the issue of Claude Desktop retriggering
oauth multiple times.

The issue of Claude Desktop running the dbt MCP twice at startup was
addressed with [this](https://github.com/dbt-labs/dbt-mcp/pull/320) PR.
This PR fixes the final time Claude Desktop starts the MCP server, when
it closes. We may decide undoing the changes made in that previous PR as
this PR fixes for both times Claude Desktop creates multiple dbt MCP
invocations.

A lot of this code is refactoring so that the credentials are provided
at a later time than the initial start of the server. This also required
propagating some async calls up the call stack. Let me know if there is
anything I can help with reviewing this.