Implement SemanticLayerClientProvider (#352)
## Summary

In [this](https://github.com/dbt-labs/dbt-mcp/pull/340) PR, I changed
how configuration worked to be lazily evaluated. In other words, the
configuration is fetched just before the API request is made.

This is problematic for our internal usage of dbt MCP, where we still
want to configure the tools when they are created rather than when they
are called. This PR adds a `SemanticLayerClientProvider` protocol such
that the behavior remains the same in this OSS repo, but we can
implement a separate `SemanticLayerClientProvider` to create the SL
client ahead of time, not lazily, as was done before PR #340. I tested
this change with `task client`.