Allow doc files to skip changie requirements (#358)
## Summary
This ignores two types of files for the `check-changelog` job.  

* `.md` files in root ONLY
* all `docs/` files

## What Changed
those files are now ignored from the changie requirement

## Why
Cleaner changelogs than a bunch of `updated docs`

* `.md` files in root only because prompts are all .md and didn't want
to have a bunch of rules
* `docs/` because they're docs

## Related Issues
<!-- Link any related issues using #issue_number -->
Closes #247 


## Checklist
- [x] I have performed a self-review of my code
- [ ] I have made corresponding changes to the documentation (in
https://github.com/dbt-labs/docs.getdbt.com) if required -- Mention it
here
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes

## Additional Notes
Not sure how you do your GHActions tests.  But in my fork, 

*
[Here](https://github.com/VDFaller/dbt-mcp/actions/runs/17989549143/job/51176197655?pr=2)
is it ignoring a doc and .md change
*
[Here](https://github.com/VDFaller/dbt-mcp/actions/runs/17989562133/job/51176241756?pr=3)
is it not ignoring a change to a prompt in src.

Co-authored-by: Devon Fulcher <24593113+DevonFulcher@users.noreply.github.com>